Tag Expiration Dates / Terminate When Past Due

+++ ec2-tag-audit/index.py
@@ -10,8 +10,9 @@ will set tags on first offense and send reminder, then will destroy
 instances if they continue to be noncompliant.
 """
 from __future__ import print_function
-from datetime import date
+from datetime import datetime, timedelta
 import fnmatch
+import json
 import time
 import boto3
 
@@ -31,35 +32,73 @@ def handler(event, context):
 
     # audit resources
     noncompliant_instances = []
-    for instance in get_instance_details():
+    for instance in get_instance_details(client):
         if not audit_tags(instance['tags'], required_tags):
-            noncompliant_instances.append(instance['id'])
+            noncompliant_instances.append(instance)
 
-    # TODO:
-    #   if node first offense, tag with expiration date (or something similar)
-    #   if expiration date has past, then terminate it
-    #
-    # IDEAS:
-    #   add DynamoDB to store terminations?
-    #   add SNS topic to notify owner's of terminated nodes?
+    # build resources to set expiration
+    # build resources to set termination
+    to_terminate = []
+    to_set_expiration = []
+    to_terminate_soon = []
+    for instance in noncompliant_instances:
+        print(instance)
+        if 'tags' in instance:
+            tag_keys = [t['Key'] for t in instance['tags']]
+            if node_first_offense(tag_keys):
+                to_set_expiration.append(instance['id'])
+            else:
+                for tag in instance['tags']:
+                    if tag['Key'] == 'expiration':
+                        expiration = tag['Value']
+                        break
 
-    # terminate nodes that are not compliant
-    client.terminate_instances(
-        InstanceIds=noncompliant_instances
-    )
+                if node_past_due(expiration):
+                    to_terminate.append(instance['id'])
+                else:
+                    to_terminate_soon.append(instance['id'])
+        else:
+            to_set_expiration.append(instance['id'])
 
-    print(noncompliant_instances)
+
+    # tag expiration dates
+    if to_set_expiration:
+        expiration_date = datetime.now() + timedelta(days=7)
+        client.create_tags(
+            Resources=to_set_expiration,
+            Tags=[
+                {
+                    'Key': 'expiration',
+                    'Value': expiration_date.strftime("%Y-%m-%d %H:%M:%S")
+                }
+            ]
+        )
+
+    # terminate instances that aren't compliant
+    if to_terminate:
+        client.terminate_instances(
+            InstanceIds=to_terminate
+        )
+
+    # TODO: remove expiration tags from items that now comply
+
+    print("Noncompliant: " + json.dumps(noncompliant_instances))
+    print("Set Expiration Date: " + json.dumps(to_set_expiration))
+    print("To Terminate Soon: " + json.dumps(to_terminate_soon))
+    print("To Terminate: " + json.dumps(to_terminate))
+
+    #print(noncompliant_instances)
     print("Function Complete")
 
 
-def get_instance_details():
+def get_instance_details(ec2_client):
     """ Get Instance Data
 
     Retrieve and compile just the data we need from EC2 instances
     """
 
     # find all instances that are not terminated
-    response = client.describe_instances(
+    response = ec2_client.describe_instances(
         Filters=[
             {
                 'Name': 'instance-state-name',
@@ -127,11 +166,32 @@ def validate_tag_value(given_value, accepted_values):
 def validate_tag_keys(given_tag_keys, required_tag_keys):
     """Validate Set Tags
 
-    Audits the keys provided by the resource and compares them to 
-    the required tags. 
+    Audits the keys provided by the resource and compares them to
+    the required tags.
 
-    given_tag_keys: list 
+    given_tag_keys: list
 
     required_tag_keys: list
     """
     return required_tag_keys == list(set(given_tag_keys).intersection(required_tag_keys))
+
+
+def node_first_offense(given_tags):
+    """Node First Offense
+
+    If expiration date hasn't been previously set, retrun True
+    """
+    return 'expiration' not in given_tags
+
+
+def node_past_due(expiration_date):
+    """Node Past Due
+
+    If an expiration date is set on a node, is it past due.
+    If it is past due return true
+
+    Timestamp example: 2016-12-22 09:44:41
+    """
+    return datetime.now() > datetime.strptime(expiration_date, "%Y-%m-%d %H:%M:%S")
+
+
